chore: move and clear the TODO for Skin Layers
